requirements.txt
google-api-python-client>=1.7.10
google-auth-httplib2>=0.0.3
google-auth-oauthlib>=0.4.0
XLsxWriter>=0.9.8
xlrd>=1.1.0
xlwt>=1.2.0
bottleneck>=1.2.1
numexpr>=2.6.2
pandas>=0.25.0
